changed code for new structure

+++ order-management-service/src/main/java/com/witchshop/ordermanagement/entity/NewOrder.java
@@ -10,6 +10,6 @@ import java.util.UUID;
 @NoArgsConstructor
 @AllArgsConstructor
 public class NewOrder {
-    private Integer pipelineId;
+    private Long pipelineId;
     private UUID clientId;
 }
