+++ order-management-service/src/main/java/com/witchshop/ordermanagement/service/DBService.java
@@ -5,7 +5,6 @@ import com.witchshop.ordermanagement.entity.OrderCreationResult;
 import com.witchshop.ordermanagement.enums.Specialization;
 import org.springframework.stereotype.Service;
 
-import java.math.BigInteger;
 import java.util.*;
 
 @Service
@@ -24,7 +23,7 @@ public class DBService {
                 requirements
         );
     }
-    public Optional<Integer> getNextStep(BigInteger pipelineId, BigInteger stepId){
+    public Optional<Integer> getNextStep(Long pipelineId, Integer stepId){
         //TODO Сначала, конечно, идёт в БД, чтобы понять, есть ли следующий шаг и кидает Optional.of(value)
         return Optional.empty();
     }
